adjusted docstrings according to mkdocs conventions for Elicit obj
